Handle the case /etc/hostid is being wrongly marked as conffile
There are packages distributed with /etc/hostid file installed as conffile,
which is going to cause problem when the package is purged. Here we detect
such case and migrate it to an ordinary file that preserves during purge.

Also replace the endianness detection code to use od(1) from coreutils
instead of requiring gcc